Organize code and add starter rake task
This loosely structures the code like a gem, although the broker is not
a gem. In addition, Goliath is prevented from auto-starting so we can
instrument things ourselves and start it programatically as necessary.

A rake task is added to make starting the server easier, usage:
`rake broker:start`. Not sure if it will stick around forever once we
add Capistrano, but it's handy for now.