Add spec infrastructure + basic test
This adds rspec/Guard to the project, as well as a basic passing test.
Testing this is a huge pain, as it's tough to avoid stubbing out a
million third-party dependencies/calls, etc.  In addition, testing
around EventMachine is weird, I'm not sure I have it fully figured out
yet.