Convert Broker to module and move Channels to SocketServer
The Broker is never actually instantiated as a class, so we can clean
things up by turning it into a module instead.

Also, Channels are only needed within the SocketServer so they are moved
there.

A remaining TODO is to set up infrastructure in the SocketServer to
handle incoming AMQP messages (using the existing subscription in the
broker)