Add strong_parameters
We'll need to validate message from Invoca's API somehow, and
strong_parameters from Rails is a great way to do that.

Adding the gem was strangely difficult, and caused a number of
dependency resolution issues. For now, it appears that requiring
strong_parameters __before__ anything else works, but after causes a
load error. Not sure why, hopefully this won't come back to bite us
later.