Add dependency to forked version of StrongParameters
Getting Strong Parameters to work outside of Rails turns out to be a
pretty painful exercise. Unbeknownst to me, the rails/strong_parameters
gem appears solely to exist as a shim for Rails 3 projects, and not as
an extraction of the functionality. I ended up forking strong_parameters,
and pasting in the code currently in the Rails 4 source tree. This is
an exercise in pain and not at all a nice solution, but at least it
works for now.