Fix Goliath hackery madness
Commit #1aa1ba5 added in some gnarly hacks in an attempt to get a handle
to the Goliath server object. This was needed because the Broker-Invoca
and Broker-Client interfaces used class/instance methods respectively
and it was difficult to bridge those.

On reflection, there was no need to use class methods at all. It was
done that way because the AMQP queue subscription was being set in the
Broker, as opposed to the socket server. This commit modifies the code
to set up the subscription at the Goliath instance level, so when
messages come in we can call all the instance methods we please.