More robust event routing
This solidifies the Invoca-Broker and Broker-Client interactions and
event handlers a little more. There's more to be done, but this is the
gist of it.

The code is all one giant lump in socket_server.rb, which makes me sad.
It'd be nice to split everything into modules at the very least.